auto.bas loader - work in progress
So the "auto.bas" already included was somewhat older version, not
really loading anything. And it was binary file, hard to edit on PC.

So the loader should be now edited as Tools/auto.bas.txt file (i.e. make
this one most recent and complete), and then tool "bas2tap v2.6" is used
to convert it to release/auto.tap file.

How to convert auto.tap into auto.bas on PC side is not clear to me,
it's TODO.

Also the loader itself, not sure of "LOAD a$" is what David was using to
load the snapshots, must verify with him that it actually works as
intended.

+++ Tools/dorelease.sh
@@ -35,7 +35,8 @@ done
 [[ $updatedSomeFile ]] && date -u +"%F %T %Z" > "release/!!built"
 # copy the readme every time
 cp release-README.txt "release/!!README.txt"
-# copy the basic loader every time
-cp Tools/auto.bas release/auto.bas
+# compile/copy the basic loader every time, if available
+command -v bas2tap 1>/dev/null && bas2tap -q -a1 -s"auto.bas" Tools/auto.bas.txt release/auto.tap
+[[ -s Tools/auto.bas ]] && cp Tools/auto.bas release/auto.bas
 # check if some file got leftovers (there should be "3" files of same base name at most)
 source Tools/chkReleaseAbundance.sh
